# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "AI Agent configuration platform with Skills and MCP support",
        "title": "Eino Agent API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@aiops.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/skills": {
            "get": {
                "description": "Get a list of all available skills",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "skills"
                ],
                "summary": "List all skills",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_fisk086_taskmate_internal_schema.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_fisk086_taskmate_internal_schema.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new skill with the specified configuration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "skills"
                ],
                "summary": "Create a new skill",
                "parameters": [
                    {
                        "description": "Skill data",
                        "name": "skill",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_fisk086_taskmate_internal_schema.CreateSkillRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_fisk086_taskmate_internal_schema.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_fisk086_taskmate_internal_schema.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_fisk086_taskmate_internal_schema.APIResponse"
                        }
                    }
                }
            }
        },
        "/skills/{id}": {
            "put": {
                "description": "Partial update: only send fields to change (risk_level, execution_mode, prompt_hint, etc.)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "skills"
                ],
                "summary": "Update a skill",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Skill ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_fisk086_taskmate_internal_schema.UpdateSkillRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_fisk086_taskmate_internal_schema.APIResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a skill by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "skills"
                ],
                "summary": "Delete a skill",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Skill ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_fisk086_taskmate_internal_schema.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_fisk086_taskmate_internal_schema.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_fisk086_taskmate_internal_schema.APIResponse"
                        }
                    }
                }
            }
        },
        "/workflows": {
            "get": {
                "description": "Configurable multi-agent orchestration templates (for UI and POST /chat with workflow_id).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workflows"
                ],
                "summary": "List agent workflows",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_fisk086_taskmate_internal_schema.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_fisk086_taskmate_internal_schema.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "kind: single | sequential | parallel | supervisor | loop (runtime executes sequential today; others reserved for Eino ADK).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workflows"
                ],
                "summary": "Create workflow",
                "parameters": [
                    {
                        "description": "Workflow",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_fisk086_taskmate_internal_schema.CreateWorkflowRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_fisk086_taskmate_internal_schema.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_fisk086_taskmate_internal_schema.APIResponse"
                        }
                    }
                }
            }
        },
        "/workflows/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workflows"
                ],
                "summary": "Get workflow by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Workflow ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_fisk086_taskmate_internal_schema.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_fisk086_taskmate_internal_schema.APIResponse"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workflows"
                ],
                "summary": "Update workflow",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Workflow ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_fisk086_taskmate_internal_schema.UpdateWorkflowRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_fisk086_taskmate_internal_schema.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_fisk086_taskmate_internal_schema.APIResponse"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "workflows"
                ],
                "summary": "Delete workflow",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Workflow ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_fisk086_taskmate_internal_schema.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_fisk086_taskmate_internal_schema.APIResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_fisk086_taskmate_internal_schema.APIResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "github_com_fisk086_taskmate_internal_schema.CreateSkillRequest": {
            "type": "object",
            "required": [
                "key",
                "name",
                "source_ref"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "source_ref": {
                    "type": "string"
                }
            }
        },
        "github_com_fisk086_taskmate_internal_schema.CreateWorkflowRequest": {
            "type": "object",
            "required": [
                "key",
                "kind",
                "name",
                "step_agent_ids"
            ],
            "properties": {
                "config": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "description": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "key": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                },
                "kind": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 200,
                    "minLength": 1
                },
                "step_agent_ids": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "github_com_fisk086_taskmate_internal_schema.UpdateSkillRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "execution_mode": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "prompt_hint": {
                    "type": "string"
                },
                "risk_level": {
                    "type": "string"
                },
                "source_ref": {
                    "type": "string"
                }
            }
        },
        "github_com_fisk086_taskmate_internal_schema.UpdateWorkflowRequest": {
            "type": "object",
            "properties": {
                "config": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "description": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "kind": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "step_agent_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT access token (Authorization: Bearer plus space plus token)",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}